requirements.txt
tinkoff-invest-api
pandas
numpy
matplotlib
scikit-learn
tqdm
pyarrow
fastparquet